Remove platform_channel_tizen.cc
