+++ shell/platform/tizen/BUILD.gn
@@ -106,7 +106,6 @@ template("embedder") {
       "channels/lifecycle_channel.cc",
       "channels/navigation_channel.cc",
       "channels/platform_channel.cc",
-      "channels/platform_channel_tizen.cc",
       "channels/platform_view_channel.cc",
       "channels/settings_channel.cc",
       "channels/settings_channel_tizen.cc",
